Fix some additional replex compile warning. Add a bit of debug for replex Fix an overflow bug that caused the data rate to be miscalculated on some HDTV streams, causing an A/V rate mismatch, and eventually the deadlock detector to keick in.  Fixes bug reported by bolek-mythtv-at-curl.com
git-svn-id: http://svn.mythtv.org/svn/trunk@8315 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/programs/mythtranscode/replex/mpg_common.h
@@ -47,7 +47,7 @@ typedef struct index_unit_s{
 	uint8_t  frame_off;
 	uint8_t  frame_start;
 	uint8_t  err;
-	uint16_t framesize;
+	uint32_t framesize;
 } index_unit;
 
 #define NO_ERR    0
